restructured my labelcode into seperate folder #8
