+++ webui/src/pages/bug/labels/ManageLabelsQuery.graphql
